+++ packages/framework/tests/Unit/Foundation/HyperlinksUrlPathHelpersTest.php
@@ -6,7 +6,6 @@ namespace Hyde\Framework\Testing\Unit\Foundation;
 
 use Hyde\Foundation\HydeKernel;
 use Hyde\Foundation\Kernel\Hyperlinks;
-use Hyde\Framework\Exceptions\BaseUrlNotSetException;
 use Hyde\Testing\TestCase;
 
 /**
@@ -155,14 +154,11 @@ class HyperlinksUrlPathHelpersTest extends TestCase
         $this->assertSame('http://localhost/foo/bar', $this->class->url('http://localhost/foo/bar/'));
     }
 
-    public function testQualifiedUrlThrowsExceptionWhenNoSiteUrlIsSet()
+    public function testQualifiedUrlReturnsNullWhenNoSiteUrlIsSet()
     {
         $this->withSiteUrl(null);
 
-        $this->expectException(BaseUrlNotSetException::class);
-        $this->expectExceptionMessage('No site URL has been set in config (or .env).');
-
-        $this->class->url();
+        $this->assertNull($this->class->url());
     }
 
     public function testHelperFallsBackToRelativeLinksWhenNoSiteUrlIsSet()